add tests for single scalar return from training
